plugins: create the favicons entirely in Ruby
