+++ src/_includes/head.html
@@ -27,14 +27,7 @@
 <meta name="theme-color" content="#{{ primary_color_light }}" media="(prefers-color-scheme: light)">
 <meta name="theme-color" content="#{{ primary_color_dark  }}" media="(prefers-color-scheme: dark)">
 
-<link rel="icon" type="image/png"    href="/f/{{ primary_color_light }}.png" media="(prefers-color-scheme: light)">
-<link rel="icon" type="image/x-icon" href="/f/{{ primary_color_light }}.ico" media="(prefers-color-scheme: light)">
-
-<link rel="icon" type="image/png"    href="/f/{{ primary_color_dark }}.png" media="(prefers-color-scheme: dark)">
-<link rel="icon" type="image/x-icon" href="/f/{{ primary_color_dark }}.ico" media="(prefers-color-scheme: dark)">
-
-<link rel="apple-touch-icon" href="/theme/apple-touch-icon.png">
-
+{% include head/icons.html %}
 {% include head/preload.html %}
 {% include head/socialgraph.html %}
