+++ src/_includes/head/icons.html
@@ -1,0 +1,44 @@
+<link
+  rel="icon"
+  type="image/png"
+  sizes="32x32"
+  href="/f/{{ primary_color_light }}-32x32.png"
+  media="(prefers-color-scheme: light)"
+>
+<link
+  rel="icon"
+  type="image/png"
+  sizes="16x16"
+  href="/f/{{ primary_color_light }}-16x16.png"
+  media="(prefers-color-scheme: light)"
+>
+
+<link
+  rel="icon"
+  type="image/png"
+  sizes="32x32"
+  href="/f/{{ primary_color_dark }}-32x32.png"
+  media="(prefers-color-scheme: dark)"
+>
+<link
+  rel="icon"
+  type="image/png"
+  sizes="16x16"
+  href="/f/{{ primary_color_dark }}-16x16.png"
+  media="(prefers-color-scheme: dark)"
+>
+
+<link
+  rel="icon"
+  type="image/x-icon"
+  href="/f/{{ primary_color_light }}.ico"
+  media="(prefers-color-scheme: light)"
+>
+<link
+  rel="icon"
+  type="image/x-icon"
+  href="/f/{{ primary_color_dark }}.ico"
+  media="(prefers-color-scheme: dark)"
+>
+
+<link rel="apple-touch-icon" href="/theme/apple-touch-icon.png">
